requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==3.2.10
django-jalali==4.3.0
django-jalali-date==1.0.1
django-redis==5.2.0
django-redis-cache==3.0.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
importlib-metadata==4.8.3
jdatetime==3.6.4
packaging==21.3
pyparsing==3.0.6
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
typing-extensions==4.0.1
wrapt==1.13.3
zipp==3.6.0
